Adding godoc comments
